UIMA-724
update match group number matching

https://issues.apache.org/jira/browse/UIMA-724

git-svn-id: https://svn.apache.org/repos/asf/incubator/uima/sandbox/trunk@615895 13f79535-47bb-0310-9956-ffa450edef68

+++ RegularExpressionAnnotator/src/main/java/org/apache/uima/annotator/regex/impl/ConceptFileParser_impl.java
@@ -218,10 +218,10 @@ public class ConceptFileParser_impl implements ConceptFileParser {
          for (int a = 0; a < annotationList.length; a++) {
 
             // create annotation position objects
-            int beginMatchGroup = annotationList[a].getBegin().getGroup();
+            int beginMatchGroup = annotationList[a].getBegin().getGroup().intValue();
             int beginLocation = annotationList[a].getBegin().getLocation()
                   .intValue();
-            int endMatchGroup = annotationList[a].getEnd().getGroup();
+            int endMatchGroup = annotationList[a].getEnd().getGroup().intValue();
             int endLocation = annotationList[a].getEnd().getLocation()
                   .intValue();
